MASAC first trial

+++ Buffer.py
@@ -11,14 +11,14 @@ class Buffer:
         self.act_dim = act_dim
 
         self.obs = np.zeros((capacity, obs_dim), dtype=np.float32)
-        self.action = np.zeros((capacity, act_dim), dtype=np.int64)
+        self.action = np.zeros((capacity, act_dim), dtype=np.float32)
         self.reward = np.zeros(capacity, dtype=np.float32)
         self.next_obs = np.zeros((capacity, obs_dim), dtype=np.float32)
         self.done = np.zeros(capacity, dtype=bool)
         """ tar_act 추가요 0723 """
-        self.tar_ent = np.zeros((capacity))
+        self.tar_ent = np.zeros(capacity, dtype=np.float32) # np.zeros((capacity))로 되어 있었는데 차이점이 없을 것 같다? 오히려 벗겨내보자.
         """ prob_act 추가요 0802 """
-        self.act_log_prob = np.zeros((capacity))
+        self.act_log_prob = np.zeros(capacity, dtype=np.float32)
 
         self._index = 0
         self._size = 0
@@ -57,13 +57,12 @@ class Buffer:
         # NOTE that `obs`, `action`, `next_obs` will be passed to network(nn.Module),
         # so the first dimension should be `batch_size`
         obs = torch.from_numpy(obs).float().to(self.device)  # torch.Size([batch_size, state_dim])
-        action = torch.from_numpy(action).int().to(self.device)  # torch.Size([batch_size, action_dim]) # 정수형만 나오도록 수정함 0804
+        action = torch.from_numpy(action).float().to(self.device)  # torch.Size([batch_size, action_dim]) # 정수형만 나오도록 수정함 0804
         reward = torch.from_numpy(reward).float().to(self.device)  # just a tensor with length: batch_size
-        # reward = (reward - reward.mean()) / (reward.std() + 1e-7)
         next_obs = torch.from_numpy(next_obs).float().to(self.device)  # Size([batch_size, state_dim])
         done = torch.from_numpy(done).float().to(self.device)  # just a tensor with length: batch_size
         """ tar_act 추가요 0723 """
-        tar_ent = torch.from_numpy(tar_ent).int().to(self.device)
+        tar_ent = torch.from_numpy(tar_ent).float().to(self.device)
         """ prob_act 추가요 0804 """
         prob_act = torch.from_numpy(prob_act).float().to(self.device)
